modify: delete config directory and move contents into newly created emailController file
